requirements.txt
aiofiles==23.1.0
aiogram==3.1.0
aiohttp==3.8.5
aiosignal==1.2.0
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.2
attrs==22.1.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.6
contourpy==1.0.5
cryptography==41.0.3
cycler==0.11.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fonttools==4.25.0
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
magic-filter==1.0.11
matplotlib==3.7.2
multidict==6.0.2
munkres==1.1.4
numexpr==2.8.4
numpy==1.25.2
openai==0.27.4
openpyxl==3.0.10
packaging==23.1
pandas==2.0.3
pandas-stubs==1.5.3.230203
Pillow==9.4.0
pip==23.2.1
plotly==5.9.0
ply==3.11
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==20.5
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
setuptools==68.0.0
sip==6.6.2
six==1.16.0
sniffio==1.3.0
tenacity==8.2.2
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.3.2
tqdm==4.65.0
types-pytz==2022.4.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wheel==0.38.4
win-inet-pton==1.1.0
yarl==1.8.1
